Refactoring: Trying to generalise the functions used
