+++ src/Components/Toolbar/Toolbar.js
@@ -2,9 +2,6 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import {
-    InputGroup,
-    InputGroupText,
-    TextInput,
     Toolbar,
     ToolbarContent as PFToolbarContent,
     ToolbarFilter,
@@ -18,11 +15,16 @@ import {
 } from '@patternfly/react-core';
 import {
     FilterIcon,
-    CalendarAltIcon,
     QuestionCircleIcon
 } from '@patternfly/react-icons';
 
 import CategoryDropdown from './CategoryDropdown';
+import CustomDateSelector from './CustomDateSelector';
+import {
+    handleCheckboxChips,
+    handleSingleChips,
+    camelToSentence
+} from './helpers';
 
 const ToolbarGroup = styled(PFToolbarGroup)`
   button {
@@ -50,52 +52,6 @@ const Select = styled(PFSelect)`
         overflow: auto;
     }
 `;
-
-/**
- * Get comparator values if their key is in the item list
- */
-export const handleCheckboxChips = (item, comparator) => {
-    if (item && comparator) {
-        return item.reduce((acc, i) => {
-            Number.isInteger(parseInt(i)) ? (i = parseInt(i)) : i;
-
-            comparator.forEach(cmpItem => {
-                if (cmpItem.key === i) {
-                    acc.push(cmpItem.value);
-                }
-            });
-
-            return acc;
-        }, []);
-    }
-
-    return [];
-};
-
-/**
- * Convert a list of objects to a list of the last value if defined
- */
-export const handleSingleChips = (date, comparator) => {
-    if (date && typeof date === 'string' && comparator) {
-        let val;
-        comparator.forEach(i => {
-            if (i.key === date) {
-                val = i.value;
-            }
-        });
-
-        if (val !== undefined) {
-            return [ val ];
-        }
-    }
-
-    return [];
-};
-
-const camelToSentence = str => {
-    const result = str.replace(/([A-Z])/g, ' $1').toLowerCase();
-    return result.charAt(0).toUpperCase() + result.slice(1);
-};
 
 const optionsForCategories = {
     status: {
@@ -137,34 +93,27 @@ const FilterableToolbar = ({
     const [ expandedItems, setExpandedItems ] = useState({});
     const [ currentCategory, setCurrentCategory ] = useState('status');
 
-    const handleStartDate = e => {
+    const onInputChange = (type, value) => {
         setFilters({
             ...filters,
-            startDate: e
-        });
-    };
-
-    const handleEndDate = e => {
-        setFilters({
-            ...filters,
-            endDate: e
+            [type]: value
         });
     };
 
     const buildFilterDropdown = () => {
+        const onToggle = type => {
+            setExpandedItems({
+                ...expandedItems,
+                [type]: !expandedItems[type]
+            });
+        };
+
         const onSelectCheckbox = (type, event, selection) => {
             setFilters({
                 ...filters,
                 [type]: event.target.checked
                     ? [ ...filters[type], selection ]
                     : filters[type].filter(value => value !== selection)
-            });
-        };
-
-        const onToggle = type => {
-            setExpandedItems({
-                ...expandedItems,
-                [type]: !expandedItems[type]
             });
         };
 
@@ -257,35 +206,12 @@ const FilterableToolbar = ({
                             }
                         />
                         { buildFilterDropdown() }
-                        { (currentCategory === 'Date' && filters.date === 'custom') && (
-                            <>
-                              <InputGroup>
-                                  <InputGroupText component="label" htmlFor="startDate">
-                                      <CalendarAltIcon />
-                                  </InputGroupText>
-                                  <TextInput
-                                      name="startDate"
-                                      id="startDate"
-                                      type="date"
-                                      aria-label="Start Date"
-                                      value={ filters.startDate || '' }
-                                      onChange={ e => handleStartDate(e) }
-                                  />
-                              </InputGroup>
-                              <InputGroup>
-                                  <InputGroupText component="label" htmlFor="endDate">
-                                      <CalendarAltIcon />
-                                  </InputGroupText>
-                                  <TextInput
-                                      name="endDate"
-                                      id="endDate"
-                                      type="date"
-                                      aria-label="End Date"
-                                      value={ filters.endDate || '' }
-                                      onChange={ e => handleEndDate(e) }
-                                  />
-                              </InputGroup>
-                            </>
+                        { (currentCategory === 'date' && filters.date === 'custom') && (
+                            <CustomDateSelector
+                                startDate={ filters.startDate }
+                                endDate={ filters.endDate }
+                                onInputChange={ onInputChange }
+                            />
                         ) }
                     </ToolbarGroup>
                 </ToolbarToggleGroup>
@@ -307,9 +233,9 @@ const FilterableToolbar = ({
                         content={
                             <div>
                                 { ' ' }
-                If enabled, nested workflows and jobs will not be included in
-                the overall totals. Enable this option to filter out duplicate
-                entries.
+                                If enabled, nested workflows and jobs will not be included in
+                                the overall totals. Enable this option to filter out duplicate
+                                entries.
                             </div>
                         }
                     >
